refactor(p1): extract ChatInputSandbox.kt Stage 1 — 21 helpers
Move 21 helpers from ChatInput.kt (1811 → 1481) to new ChatInputSandbox.kt:
- URL helpers: normalizedWebPreviewUrl, firstHttpUrl, webHostPreview, compactForSandbox, asValidThumbnailFile
- WebView ops: extractReadablePage, scheduleReadableExtracts, scheduleThumbnailCaptures, captureWebViewThumbnail, disablePreviewDarkening
- Bitmap: looksBlank
- WebViewOperationState ext: matchesPreview, bestThumbnailFile
- Sandbox status: sandboxStatusLabel/Container/OnContainerColor
- UiState ext props: operationPreviewKind/Text/Url/stepProgressText/terminalTranscript
- HTTP_URL_REGEX (file-private in new file)

All moved helpers promoted to `internal` so remaining composables in ChatInput.kt can call them. Drop 9 orphan imports (Bitmap, Canvas, Build, WebSettings, AndroidWebView, WebViewLink, JSONArray, JSONObject, URLEncoder) made dead by this extraction. Preserve byte-equal helper bodies. Baseline 6/427 unit tests preserved.

Stage 2 (22 composables + SandboxPeekBar) and Stage 3 (SandboxActivitySheet cross-pkg) to follow.
